refactor(heartbeat): move heartbeat prompt builder into bounded module
